chore: profile component
